+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+## 0.5.1 (2017-11-09)
+
+* TimeSpent on stop call
+
+## 0.5.0 (2017-09-25)
+
+* PlayerID from videojs
+* Deleted ConcurrenceViewIdMaker
+
 ## 0.4.8 (2017-08-24)
 
 * Added option showAlert (default: true)